backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.5.4
requests==2.31.0
beautifulsoup4==4.12.3
pymupdf==1.24.3
python-dotenv==1.0.1
# Optional: set SERPAPI_KEY in .env for reliable Google Jobs search (100 free/month)
# google-search-results==2.4.2
